analysis

+++ FTAnalysis/src/main/scala/LossObservation.scala
@@ -1,5 +1,5 @@
 import de.unijena.bioinf.ChemistryBase.chem.MolecularFormula
 
-case class LossObservation(loss:MolecularFormula, count:Int, frequency:Double) {
+case class LossObservation(loss:MolecularFormula, count:Double, frequency:Double) {
 
 }
